refactor: Dealer와 Player의 공통 속성과 행위를 Gamer로 추상화 한 뒤 상속하도록 수정
