+++ src/main/java/domain/dealer/Dealer.java
@@ -3,27 +3,25 @@ package domain.dealer;
 import domain.card.Card;
 import domain.card.CardBundle;
 import domain.card.CardDeck;
+import domain.player.Gamer;
 import domain.player.Player;
 import domain.player.PlayerName;
 
 import java.util.List;
 import java.util.stream.Stream;
 
-public class Dealer {
+public class Dealer extends Gamer {
 
     private static final String CANNOT_HAND_OUT_CARDS_ERROR = "딜러가 카드를 나눠줄 수 없습니다.";
-    private static final String DEFAULT_DEALER_NAME = "딜러"; // NOTE Player와 추상화를 위해서는 수정 필요
+    private static final String DEFAULT_DEALER_NAME = "딜러";
     private static final int ADDITIONAL_DRAW_CONDITION = 16;
     private static final int INITIAL_DEAL_COUNT = 2;
     private static final int HIT_COUNT = 1;
 
-    private final PlayerName name;
     private final CardDeck cardDeck;
-    private CardBundle cardBundle;
 
     private Dealer(PlayerName name, CardDeck cardDeck) {
-        this.name = name;
-        this.cardBundle = CardBundle.empty();
+        super(name);
         this.cardDeck = cardDeck;
     }
 
@@ -43,12 +41,38 @@ public class Dealer {
         return handOutCardToPlayer(player, HIT_COUNT);
     }
 
+    public CardBundle dealMyself() {
+        cardBundle = drawMySelf(INITIAL_DEAL_COUNT);
+        return cardBundle;
+    }
+
+    public CardBundle hitMyself() {
+        cardBundle = drawMySelf(HIT_COUNT);
+        return cardBundle;
+    }
+
+    public boolean hitIfRequired() {
+        if (canHit()) {
+            hitMyself();
+            return true;
+        }
+        return false;
+    }
+
+    private boolean canHit() {
+        return cardBundle.getBasicScore() <= ADDITIONAL_DRAW_CONDITION;
+    }
+
+    private CardBundle drawMySelf(int tryCount) {
+        return cardBundle.add(handOutCard(tryCount));
+    }
+
     private CardBundle handOutCardToPlayer(Player player, int tryCount) {
         CardBundle cardBundle = handOutCard(tryCount);
         return player.addCardBundle(cardBundle);
     }
 
-    public CardBundle handOutCard(int tryCount) {
+    private CardBundle handOutCard(int tryCount) {
         try {
             List<Card> cardList = Stream.generate(this::drawCard)
                     .limit(tryCount)
@@ -59,33 +83,8 @@ public class Dealer {
         }
     }
 
-    public boolean hitIfRequired() {
-        if (canHit()) {
-            drawMySelf(HIT_COUNT);
-            return true;
-        }
-        return false;
-    }
-
-    private boolean canHit() {
-        return cardBundle.getBasicScore() <= ADDITIONAL_DRAW_CONDITION;
-    }
-
-    public CardBundle dealMyself() {
-        cardBundle = drawMySelf(INITIAL_DEAL_COUNT);
-        return cardBundle;
-    }
-
-    private CardBundle drawMySelf(int tryCount) {
-        return cardBundle.add(handOutCard(tryCount));
-    }
-
-    public Card drawCard() {
+    private Card drawCard() {
         return cardDeck.giveCard();
-    }
-
-    public int getResultScore() {
-        return cardBundle.getResultScore();
     }
 
     public String toDisplayMyName() {
@@ -94,10 +93,6 @@ public class Dealer {
 
     public List<String> disPlayMyCardBundle() {
         return cardBundle.toDisplay();
-    }
-
-    public boolean isBusted() {
-        return cardBundle.isBusted();
     }
 
 }
